Day07 fix compilation + default to clang
Turns out I was using 'c++' which is ccache apparently on Fedora. That
compiled no problems (g++?) but clangd was complaining about the way I
was defining stuff. Not sure I understand but switching around the order
and forgetting about the forward declarations worked in the header.

Switched to clang++ as the 'adv' script's C++ builder just to stay
consistent.